refactor: split contentful fetch to its own file
